Added the beginnings of a login/registration system
refs #7

The first thing is to add a userid column to the database.  Yeah, forgot
that.  Sort of necessary for users

We also added were some utility functions to the Application
Controller.  These allow us to check to see if someone is logged in, and
if not, redirect to the login url.  We also added a login route to the
routes.

In the users controller, add login and logout actions.  Only login is
implemented, which redirects to CAS for the login.

Also, this commit fixes up the functional tests for the users controller
which were previously broken.