Force new users to fill out a user form (closes #8)
This commit forces new users to go to the new user form and fill it out.
The userid is assigned from the session, not from the form, to prevent
malicious userids.  The edit view/action are similarly edited to provide
update ability, while not allowing the change of the userid.