Implements feature #10: "A user should be able to logout" http://develop.feds.ca/issues/10
Adds and implements UsersController#logout, which clears
the session and redirects to CAS logout.

Makes current_user and logged_in? available in the view
(via ApplicationHelper).

Adds and implements User#to_s for a standard friendly
representation (full name).

Adds an appropriate login/logout link to the application
layout (depending on whether the user is logged in).

UsersController#login now accepts a return_to parameter
so that user-initiated login via a login link (as opposed
to forced login via login_required) redirects back to the
page the user was viewing. A newly-added helper method,
login_redirect_path, generates an appropriate URI for this.

(closes #10)

Signed-off-by: Eric Gerlach <egerlach@feds.uwaterloo.ca>