Implements feature #11: "Users should have Roles" http://develop.feds.ca/issues/11
Adds the Role model. A user may have many roles (and
conversely, a role has many users). This will be used to
implement authorization in combination with the upcoming
Permission model and related code.

Adds Role scaffolding (RolesController and its views).

Moves user new/edit form into a partial and added checkboxes
to manage roles. Currently anyone can put anyone in any
role, but this will be fixed when authorization is implemented.

Changes User validation to allow blank strings to be put
in for student number and not fail validation (as expected,
this simply causes it not to be changed). The previous
version only checked for nil; whereas when the form is
submitted with no value it is assigned a value of "" --
the empty string -- which does not evaluate to false.
The blank? method, which Rails defines on strings as well
as nil, handles this.

Gives User#student_number= a more clear argument name.

Replace manual deletion from params in UsersController#update
with attr_protected on User. This also applies to other
mass-assignment (User.new, User.update_attributes, etc.) and
seems cleaner.

(closes #11)

Signed-off-by: Eric Gerlach <egerlach@feds.uwaterloo.ca>