Implements feature #12: "Roles should be able to be assigned permissions" http://develop.feds.ca/issues/12
Adds the Permission model. A role may have many permissions
(and conversely, a permission may have many roles). This
will be used to implement authorization in combination with
the existing Role model and new code.

Adds Permission scaffolding (PermissionsController and its views).

Moves role new/edit form into a partial and adds checkboxes
to manage permissions. Currently anyone can assign any permission
to any role, but this will be fixed when authorization is implemented.

Adds and implements User#has_permission?, which looks up a permission
and checks whether the user has it. If given a Permission model, it
checks that permission. If given an integer, it looks up a permission
with that ID. If given a string, it looks up a permission with that
name. If given a symbol, it looks up a permission with a name matching
the humanized version of that symbol (for instance, :eat_cake corresponds
to "Eat cake").

(closes #12)

Signed-off-by: Eric Gerlach <egerlach@feds.uwaterloo.ca>