Implements feature #14 "A user should be able to C/R/U/D a trip"
Adds the Trip model, and unit and functional tests. The functional tests, ensure that only authenticated users with the trips permission are able to make any changes to the trips data.

The edit/new form was broken out into a partial _form.html.erb and the time_select helpers were formatted to only ask for the time without the date and to not include the dummy date fields in the form as it does by default.
Modified the display of the departure, arrival, and return times for the show and index views to just show the time instead of the whole date as well. ActiveRecord only supports Date and DateTime objects and not Time so we have to store these values as DateTimes even though we just want the time portion. The application helper datetime_to_strtime was added to preform the DateTime object to time string conversion; the time is formatted in 24-hour time displaying only hours and minutes separated by a colon.

The sales_lead validation accepts any number of positive days before the trip as a sales lead. Allows for Ottawa busses during reading week to sell 2 weeks in advance.
The return_trip parameter is a positive integer or blank value (this allows one way trips like the Ottawa trip) which corresponds to a relative number of days from the trip day that a return trip is from when the busses are actually created. It will be converted to a date when busses are made.

Access control was placed on the trips controller to ensure that a user is logged in with the correct permissions before proceeding with any of the edit/new/delete functions. Any user including an unauthenticated one can list all or display a single trip.

Added a parse_post helper function to the trips controller. This is used because the form time helpers use two variables for each time, one for the hour and another for minutes while the model expects the post to return a hash where the time is a single value.

Adds the validates_timeliness gem (http://github.com/adzap/validates_timeliness) for use when validating the departure, arrival and return times of a Trip.
NOTE: This plugin does not validate 12 hour times as you would expect. 13am is considered a valid time for instance. See the open issue: Unexpected Behavior: validates_time Accepts Invalid 12 Hour Times (http://github.com/adzap/validates_timeliness/issues#issue/11)

Test helper functions were created for less repetitive testing of valid and invalid value assignment to model properties. The two functions are assign_valid_value and assign_invalid_value, they accept a model, a model attribute, and a test value as parameters. Each function tests if the model.save method fails or there is an ActiveRecord::Validations error on the model attribute supplied.
Added the setup_authenticated_user_with_permission functional test helper function that accepts a user and permission symbol that corresponds to a user and permission fixture and fakes an authenticated user with that permission for the request being tested.

closes #14

Signed-off-by: Eric Gerlach <egerlach@feds.uwaterloo.ca>