Closes #30: Seed default authorization information into the database.
Seeds default roles and permissions into the database in db/seeds.rb
(can be executed with 'rake db:seed').

Locked down access control functionality to authorized users only.

Enhanced documentation in an RDoc-friendly way (build documentation
using 'rake doc:app', view doc/app/index.html to see the results).

Made the session secret dynamically generated the first time the
environment is loaded (and appropriately updated .gitignore) so that
the session secret in actual use does not need to be publicly
disclosed in the git repository.

Added uniqueness constraints and database indices to guarantee
expected behaviour for logic which does lookups based on columns
such as name (in permissions or roles) or userid (in users), as
well as improving performance for lookups in a number of columns.

Added 'fedbus' Rake namespace, which includes the following tasks
(as well as some other tasks used by fedbus:setup, though they will
probably not need to be called directly):

* fedbus:admin -- Adds a new user as an administrator. This makes it
  possible for someone with access to the server to add an administrator,
  a task which would be impossible from the web interface if there were
  no existing admins (or other users with the "Manage access control"
  permission).

* fedbus:setup -- Used for initial setup of the application. Confirms
  the database setup with the user, creates it (if possible), runs
  migrations and seeds the database. Prompts the user to create an admin
  user and notifies the user about other configuration steps that may be
  needed.